Show error when find returns no result
Currently, when user uses the find command if there are no names
within the contact book, the message of 0 person listed will show and
the list will clear all contacts. This could make the user experience
bad for the user given that all the contacts are cleared and the user
has to retype the intended search parameters.

Let's
- Ensure that there are search results before applying the filter
- Throw CommandException when no matching persons are found
- Keep the current list visible instead of clearing to an empty list
- Add MESSAGE_NO_PERSON_FOUND constant in Messages
- Preserve command text in input box for easy retry

This improves UX by allowing users to easily modify their search
criteria instead of retyping the entire command. The list will
also not be cleared unnecessarily.